+++ lib/restclient/request.rb
@@ -51,17 +51,6 @@ module RestClient
     end
 
     def execute &block
-      execute_inner &block
-    rescue Redirect => e
-      @processed_headers.delete "Content-Length"
-      @processed_headers.delete "Content-Type"
-      @url = e.url
-      @method = :get
-      @payload = nil
-      execute &block
-    end
-
-    def execute_inner &block
       uri = parse_url_with_auth(url)
       transmit uri, net_http_request_class(method).new(uri.request_uri, processed_headers), payload, &block
     end
@@ -201,32 +190,20 @@ module RestClient
       http_response
     end
 
-    def process_result res
+    def process_result res, &block
       if @raw_response
         # We don't decode raw requests
-        response = RawResponse.new(@tf, res)
+        response = RawResponse.new(@tf, res, args)
       else
-        response = Response.new(Request.decode(res['content-encoding'], res.body), res)
+        response = Response.new(Request.decode(res['content-encoding'], res.body), res, args)
       end
 
-      code = res.code.to_i
-
-      if (301..303).include? code
-        url = res.header['Location']
-
-        if url !~ /^http/
-          uri = URI.parse(@url)
-          uri.path = "/#{url}".squeeze('/')
-          url = uri.to_s
-        end
-        raise Redirect, url
+      if block_given?
+        block.call response, &block
       else
-        if block_given?
-          yield response
-        else
-          response.return!
-        end
+        response.return! &block
       end
+
     end
 
     def self.decode content_encoding, body